Add Terraform configuration for the Lambda endpoint.
Add Dockerfile and requirements.txt to allow building an image to provide for the Lambda endpoint.
Update the endpoint code to add the lambda_handler wrapper.
Note that this deploys but fails on testing.  The app.py reads the model from S3 and tried to write to disk, which isn't available in lambda, or at least not how I have it configured.  Next steps are to refactor so that the model is included in the Docker image.