Add a time/date chooser to the ControlsCard.
Refactor to add a "Compute Risk" button on the ControlCard for user to request a risk assessment, rather than firing off a request once all data was ready, or when changed.
Move maintenance of route and prediction data to the top-level so that collections of routes and predictions may be shown.
Move messaging the model to the top level.
Update display of results to include risk measure in the table, and display speed limit as an integer.